ass_render: Break at last position if no space was seen
It's pretty common (in CJK texts especially) to not have any spaces to
split on. Currently, these cases simply overflow the render area.

If this happens, try a last ditch attempt to wrap at the last character
seen. This dramatically improves the usability of long subtitles with no
clear break point.